requirements.txt
# ===== LangChain core (1.x) =====
langchain>=1.1.0,<2.0.0
langchain-core>=1.1.3,<2.0.0

# ===== Community & integrations =====
langchain-community>=0.4.0,<0.5.0
langchain-openai>=1.0.0,<2.0.0
langchain-huggingface>=1.0.0,<2.0.0
langchain-google-genai
google-generativeai

# ===== Vector store =====
chromadb>=1.3.5,<2.0.0
langchain-chroma>=1.1.0,<2.0.0

# ===== Embedding =====
sentence-transformers>=2.6.0

# ===== File / text processing =====
pypdf>=4.0.0
chardet>=5.2.0
tiktoken
reportlab

# ===== UI =====
streamlit>=1.32.0

# ===== Math =====
numpy>=1.26.0

